f2f_Q3_2019/requirements.txt
amqp==2.5.0
aniso8601==6.0.0
billiard==3.6.0.0
celery==4.3.0
Django==2.2.4
djangorestframework==3.10.2
graphene==2.1.7
graphene-django==2.4.0
graphql-core==2.2.1
graphql-relay==2.0.0
kombu==4.6.3
promise==2.2.1
psycopg2-binary==2.8.3
pytz==2019.2
redis==3.3.4
requests==2.22.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
vine==1.3.0
